+++ packages/dso-toolkit/components/Componenten/dropdown-button/dropdown-button.config.js
@@ -1,5 +1,55 @@
 module.exports = {
-  notes: 'Toont een dropdown met daarin een gevulde uitklapper. Gebaseerd op de Bootstrap "Dropdowns" component. Class `dso-dropdown-right` op de parent van of op `.dropdown` zelf zorgt voor een rechter uitlijning.',
+  notes: `
+  **Implementatie:**
+
+  Het openen en sluiten van de dropdown button moet worden ondersteund met muis-, touch- en toetsenbordbediening.
+
+  De scripting zal zelf geïmplementeerd moeten worden, waarbij het te programmeren gedrag op de diverse relevante toetsen gelijk moet zijn aan de native browser dropdown/select werkvorm:
+
+  * \`event.code\` \`ArrowUp\` : omhoog, vorige 'option', zelfde als tab
+  * \`event.code\` \`ArrowDown\` : omlaag, volgende 'option', zelfde als shift-tab
+  * \`event.code\` \`Escape\` : sluit dropdown, geen keuze maken
+  * \`event.code\` \`Space\` : maak keuze
+
+
+  (inspiratie: [Bootstrap dropdown](https://github.com/twbs/bootstrap/blob/main/js/src/dropdown.js#L450))
+
+  **Voorschriften/algemeen:**
+
+  * \`button\` en \`dropdown-menu\` staan in een \`div.dropdown\`. Het \`dropdown-menu\` wordt getoond wanneer \`div.dropdown\` de class \`open\` heeft.
+  * \`button\` attribute \`aria-haspopup="true"\`
+
+
+  **Dropdown-button ingeklapt:**
+
+  * \`div.dropdown\` geen class \`open\`
+  * \`button\` attribute \`aria-expanded="false"\`
+
+
+  Markup:
+  \`\`\`
+  <div class="dropdown">
+    <button type="button" aria-haspopup="true" aria-expanded="false" class="btn btn-link">
+    <div class="dropdown-menu">
+  </div>
+  \`\`\`
+
+  **Dropdown-button ingeklapt:**
+
+  * \`div.dropdown\` heeft class \`open\`
+  * \`button\` attribute \`aria-expanded="true"\`
+
+
+  Markup:
+  \`\`\`
+  <div class="dropdown open">
+    <button type="button" aria-haspopup="true" aria-expanded="true" class="btn btn-link">
+    <div class="dropdown-menu">
+  </div>
+  \`\`\`
+
+  Note: de \`button\` kan naar behoeve de classes \`btn-link\`, \`btn-default\` en \`btn-primary\` hebben.
+  `,
   status: 'ready',
   collated: true,
   collator: function (markup, item) {